refactor: API 관련 테스트 클래스 분리 및 기존 테스트 코드 이동
